Classs renaming.
For classes that implement an interface (or multi-interfaces) and can be
used to create instances,  postfix "Impl" are removed.  Impl are only
used for classes implement a particular interface and are supposed to be
used as a base class for classes implementing that interface. In other
words, a class can derive from a Impl class to implement a certain
class.
1. DimensionsImpl renamed to Dimensions;
2. DataImpl renamed to DataObject;